[TASK] Migration path from tt_address_group to sys_category
In tt_address 3.0, the Category API is used instead of the built-in
address groups in tt_address. Therefore, a wizard is added to the
install tool that converts tt_address_group items to sys_category
items (respecting language), writes the MM relations and also
updates the group selection in the FlexForm.

Releases: 3.0
Resolves: #71430
Change-Id: I9f2bd174d4db94361e63a12d34ca05b9df0bf250
Reviewed-on: https://review.typo3.org/44641
Reviewed-by: Thomas Schlumberger <thomas@b13.de>
Tested-by: Thomas Schlumberger <thomas@b13.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>