Add link to single view in map info bubble
Add a link to the single view page if the custom typoscript setup variable settings.map.singlePid is set. 
Please see extended typoscript in setup,typoscript as well. 
Actually id would be best if the already existing typoscript setup variable settings.singlePid and the according value in the plugin would be considered for this to work, but this value doesn't work (yet) in the partial Maps.html