[BUGFIX] Cannot add new address record in TYPO3 v12 PHP 8.2 (#512)
* [BUGFIX] Implement safeguard for non-numeric UID in Label.php

The code has been adjusted to handle situations where the UID is not numeric in the Label.php class (e.g. a new record). Instead of directly casting the UID to an integer, it is first checked whether the UID is numeric. If it is not, the row is fetched directly from the parameters. This aims to prevent potential casting errors and provides more robust handling for uid parameters.

Resolves: #511

* Remove unnecessary DebuggerUtility import