[TASK] Use Connection instead of PDO
In Doctrine DBAL v4, as described in the documentation, support for using the \PDO::PARAM_* constants has been dropped in favor of the enum types. This should be migrated to Connection::PARAM_* in order to be compatible with TYPO3 v13 later on. Connection::PARAM_* can already be used now as it is compatible with TYPO3 11 and 12.